refactor: extract anecdote creation into separate AnecdoteForm component
Moved anecdote creation form and logic from App.jsx to a dedicated AnecdoteForm component. This improves modularity and adheres to the single responsibility principle by separating concerns. App.jsx now focuses solely on displaying anecdotes and handling votes, while AnecdoteForm manages anecdote creation using the createAnecdote action creator.
